changed

+++ src/main.rs
@@ -7,7 +7,7 @@ fn main() {
     let universe = mpi::initialize().unwrap();
     let world = universe.world();
     let rank = world.rank() as usize;
-    let size = world.size() as i32;
+    let size = world.size() as usize;
     let root_rank = 0i32;
     let repeat = 2 as usize;
     let root_process = world.process_at_rank(root_rank);
